backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bson==0.5.10
certifi==2023.7.22
click==8.1.6
dnspython==2.4.0
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
motor==3.2.0
pydantic==2.0.3
pydantic-core==2.3.0
pymongo==4.4.1
python-dateutil==2.8.2
python-decouple==3.8
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.1
